requirements.txt
appier==1.32.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.5.14
Django==5.0.1
facebook-sdk==3.1.0
idna==3.6
marshmallow==3.20.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
python-decouple==3.8
python-facebook-api==0.18.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
